+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2 (15 July, 2019)
  - [BC] REMOVED: Method Yaroslavche\ConfigUIBundle\Service\Config#getBundleConfigDefinitions() was removed
  - [BC] REMOVED: Method Yaroslavche\ConfigUIBundle\Service\Config#getBundles() was removed
+ - [BC] CHANGED: The number of required arguments for Yaroslavche\ConfigUIBundle\Service\Config#__construct() increased from 2 to 3
  - [BC] REMOVED: Class Yaroslavche\ConfigUIBundle\Controller\DashboardController has been deleted
 
 ## 0.0.1 (14 July, 2019)